// src/api/generated/models/marketingCampaignControllerSendBodyChannel.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type MarketingCampaignControllerSendBodyChannel = typeof MarketingCampaignControllerSendBodyChannel[keyof typeof MarketingCampaignControllerSendBodyChannel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MarketingCampaignControllerSendBodyChannel = {
  email: 'email',
  whatsapp: 'whatsapp',
} as const;